// src/modules/redis/enum/redis-prefix-enum.ts
export enum RedisPrefixEnum {
  REQUEST_USER_DATA = 'request_user_data',
  ACCESS_TOKEN = 'access_token',
  AGENCY = 'agency',
  CAMPAIGN_DETAILS = 'campaign_details',
  CAMPAIGN_VOTES = 'campaign_votes',
  VOTABLE_CONTENT = 'votable_content',
  DASHBOARD_FIND_STATUS_COUNT = 'findStatusCount',
  DASHBOARD_FIND_OVERALL_REPORT = 'findOverallReport',
  CAMPAIGN_TEMPLATES = 'campaignTemplateList',
  CAMPAIGN_USER_TEMPLATES = 'campaignUserTemplateList',
  DASHBOARD_POPULAR_CAMPIGN = 'findpopularcampaign',
  WHATS_APP_INTEGRATION = 'whats_app_integration',
  CONTACT_IMPORT_LOCK = 'CONTACT_IMPORT_LOCK', // Add this

}